requirements.txt
<<<<<<< HEAD
streamlit
arxiv
google-genai
=======
altair
pandas
streamlit
>>>>>>> 8c090af (Duplicate from streamlit/streamlit-template-space)
